img/live: use callPackage function instead of "with pkgs;"
When cross compiling Spectrum OS for aarch64 it tries to run aarch64
binaries on x64 host. To fix this issue it is possible to add
pkgsBuildHost prefix to all native build inputs. However, a better
option would probably be to use callPackage function. With that change
it is able to automatically select correct packages.

Signed-off-by: Yuri Nesterov <yuriy.nesterov@unikie.com>
Message-Id: <20220927123318.2531867-1-yuriy.nesterov@unikie.com>
Signed-off-by: Alyssa Ross <alyssa.ross@unikie.com>