img/app: add support for testing virtio-gpu
crosvm sadly doesn't support socket activation or readiness
notification, hence the sleep loop here.

I've removed the tap device as it required privileges to set up, and
was difficult to actually test with since it wouldn't be set up
correctly on the host in the same way the Spectrum host would set it
up.

Signed-off-by: Alyssa Ross <alyssa.ross@unikie.com>